refactor: reorganize components folder by functionality
Reorganize components into logical groupings:

src/components/
├── auth/                      # Authentication components
│   ├── AuthFormLayout.tsx
│   ├── PasswordInput.tsx
│   ├── ProtectedRoute.tsx     (moved from components/)
│   └── index.ts
│
├── feedback/                  # User feedback components (NEW)
│   ├── Modal/                 (moved from components/Modal/)
│   │   ├── Modal.tsx
│   │   ├── ConfirmModal.tsx
│   │   ├── AlertModal.tsx
│   │   └── index.ts
│   ├── Toast/                 (moved from components/Toast/)
│   │   ├── Toast.tsx
│   │   ├── ToastContext.tsx
│   │   └── index.ts
│   ├── SaveIndicator.tsx      (moved from components/)
│   └── index.ts
│
├── layout/                    # Layout components (NEW)
│   ├── Layout.tsx             (moved from components/)
│   ├── PageLoadingFallback.tsx    (moved)
│   ├── ComponentLoadingFallback.tsx (moved)
│   └── index.ts
│
├── financial/                 # Financial components
│   ├── FinancialMetrics.tsx
│   ├── FinancialCharts.tsx
│   └── index.ts
│
└── ui/                        # UI primitives
    ├── Button.tsx
    ├── Card.tsx
    ├── Input.tsx
    ├── Textarea.tsx
    ├── Badge.tsx
    ├── Progress.tsx
    ├── Spinner.tsx
    └── index.ts

Benefits:
- Clear functional groupings
- Easier to find related components
- Better maintainability
- Consistent barrel exports

Build Status: ✅ Successful
