+++ src/components/auth/ProtectedRoute.tsx
