+++ src/components/auth/index.ts
@@ -2,9 +2,13 @@
  * 파일명: index.ts
  * 
  * 파일 용도:
- * auth 컴포넌트 barrel export
+ * 인증 관련 컴포넌트 barrel export
+ * - 인증 폼 레이아웃
+ * - 비밀번호 입력
+ * - 보호된 라우트
  */
 
 export { PasswordInput } from './PasswordInput';
 export { AuthFormLayout } from './AuthFormLayout';
+export { ProtectedRoute } from './ProtectedRoute';
