+++ src/components/feedback/Modal/ConfirmModal.tsx
@@ -20,7 +20,7 @@
 
 import React from 'react';
 import { AlertTriangle, Info } from 'lucide-react';
-import { Modal } from '@/components/Modal/Modal';
+import { Modal } from './Modal';
 import { cn } from '@/common/utils';
 
 export interface ConfirmModalProps {
