+++ src/components/feedback/index.ts
@@ -1,0 +1,19 @@
+/**
+ * 파일명: index.ts
+ * 
+ * 파일 용도:
+ * 피드백 컴포넌트 모듈 barrel export
+ * - Toast 알림
+ * - Modal 다이얼로그
+ * - 저장 상태 표시
+ */
+
+// Toast
+export * from './Toast';
+
+// Modal
+export * from './Modal';
+
+// SaveIndicator
+export { SaveIndicator } from './SaveIndicator';
+
