+++ src/components/layout/index.ts
@@ -1,0 +1,13 @@
+/**
+ * 파일명: index.ts
+ * 
+ * 파일 용도:
+ * 레이아웃 컴포넌트 모듈 barrel export
+ * - 메인 레이아웃
+ * - 로딩 폴백 컴포넌트
+ */
+
+export { Layout } from './Layout';
+export { PageLoadingFallback } from './PageLoadingFallback';
+export { ComponentLoadingFallback } from './ComponentLoadingFallback';
+
